Transform image overrides on way in and out to ensure correct meta
